+++ tests/engine.php
@@ -59,6 +59,8 @@ class CastorTest extends \PHPUnit\Framework\TestCase {
         $tpl = new \Jelix\Castor\Castor(self::$castorConfig);
         $json = json_encode(['foo'=>'foo value', 'bar'=>'bar value']);
         $tpl->assign('myjson', $json);
+        $tpl->assign('mydate', '2024-08-25');
+
         $result = $tpl->fetch('assets/modifiers.tpl');
         $this->assertEquals(file_get_contents(__DIR__.'/assets/modifiers.txt'), $result);